New class AspectRatioSelection
Changed aspect ratio components to double
Redesigned calculate_width - moved to Selection class

+++ cinelerra/edlsession.h
@@ -81,8 +81,8 @@ public:
 	AutoConf *auto_conf;
 	float actual_frame_rate;
 // Aspect ratio for video
-	float aspect_w;
-	float aspect_h;
+	double aspect_w;
+	double aspect_h;
 	int audio_channels;
 	int audio_tracks;
 // automation follows edits during editing
